04_agents/requirements.txt
langchain==0.0.352
langchain-openai==0.0.5
python-dotenv==1.0.0
pyboxen